requirements.txt
numpy
opencv-python>=4.5.4
websockets==11.0.3
Pillow
simple-pid